create skip main hook, add to admin and rider routes
