+++ frontend/src/pages/Admin/Routes.tsx
@@ -1,11 +1,11 @@
-import React, { useEffect, useRef, useState } from 'react';
+import React, { useState } from 'react';
 import {
   BrowserRouter as Router,
   Route,
   Switch,
   Redirect,
-  useHistory,
 } from 'react-router-dom';
+import useSkipMain from '../../hooks/useSkipMain';
 import Sidebar from '../../components/Sidebar/Sidebar';
 import Employees from './Employees';
 import Students from './Students';
@@ -18,66 +18,63 @@ import ExportPreview from '../../components/ExportPreview/ExportPreview';
 import DateContext from '../../context/date';
 
 const Routes = () => {
+  const skipRef = useSkipMain();
+  return (
+    <>
+      <div tabIndex={-1} ref={skipRef}></div>
+      <a className='skip-main' href='#main'>Skip to main content</a>
+      <Sidebar type="admin">
+        <Switch>
+          <Route
+            path="/home"
+            render={({ match: { url } }) => (
+              <>
+                <Route path={`${url}/`} component={Home} exact />
+                <Route path={`${url}/export`} component={ExportPreview} />
+              </>
+            )}
+          />
+          <Route
+            path="/employees"
+            render={({ match: { url } }) => (
+              <>
+                <Route path={`${url}/`} component={Employees} exact />
+                <Route path={`${url}/employee`} component={EmployeeDetail} />
+              </>
+            )}
+          />
+          <Route
+            path="/riders"
+            render={({ match: { url } }) => (
+              <>
+                <Route path={`${url}/`} component={Students} exact />
+                <Route path={`${url}/rider`} component={RiderDetail} />
+              </>
+            )}
+          />
+          <Route path="/locations" component={Locations} />
+          <Route path="/analytics" component={Analytics} />
+          <Route path="*">
+            <Redirect to="/home" />
+          </Route>
+        </Switch>
+      </Sidebar>
+    </>
+  );
+};
+
+const AdminRoutes = () => {
   const [curDate, setCurDate] = useState(new Date());
   const defaultVal = { curDate, setCurDate };
-  const history = useHistory();
-  const ref = useRef<HTMLDivElement>(null);
-  useEffect(() => {
-    const unlisten = history.listen(() => {
-      if (ref) {
-        ref.current?.focus();
-      }
-    });
-    return unlisten;
-  }, [history]);
 
   return (
     <DateContext.Provider value={defaultVal}>
       <Router basename="/admin">
-      <>
-      <div tabIndex={-1} ref={ref}></div>
-      <a className='skip-main' href='#main'>Skip to main content</a>
-        <Sidebar type="admin">
-          <Switch>
-            <Route
-              path="/home"
-              render={({ match: { url } }) => (
-                <>
-                  <Route path={`${url}/`} component={Home} exact />
-                  <Route path={`${url}/export`} component={ExportPreview} />
-                </>
-              )}
-            />
-            <Route
-              path="/employees"
-              render={({ match: { url } }) => (
-                <>
-                  <Route path={`${url}/`} component={Employees} exact />
-                  <Route path={`${url}/employee`} component={EmployeeDetail} />
-                </>
-              )}
-            />
-            <Route
-              path="/riders"
-              render={({ match: { url } }) => (
-                <>
-                  <Route path={`${url}/`} component={Students} exact />
-                  <Route path={`${url}/rider`} component={RiderDetail} />
-                </>
-              )}
-            />
-            <Route path="/locations" component={Locations} />
-            <Route path="/analytics" component={Analytics} />
-            <Route path="*">
-              <Redirect to="/home" />
-            </Route>
-          </Switch>
-        </Sidebar>
-      </>
+        <Routes />
       </Router>
     </DateContext.Provider >
 
   );
 };
 
-export default Routes;
+export default AdminRoutes;
